+++ src/main/java/com/viglet/shiohara/utils/ShFolderUtils.java
@@ -1,11 +1,14 @@
 package com.viglet.shiohara.utils;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
+import org.apache.http.client.ClientProtocolException;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -15,12 +18,16 @@ import com.viglet.shiohara.persistence.model.folder.ShFolder;
 import com.viglet.shiohara.persistence.model.object.ShObject;
 import com.viglet.shiohara.persistence.model.post.ShPost;
 import com.viglet.shiohara.persistence.model.post.ShPostAttr;
+import com.viglet.shiohara.persistence.model.post.type.ShPostType;
 import com.viglet.shiohara.persistence.model.reference.ShReference;
 import com.viglet.shiohara.persistence.model.site.ShSite;
 import com.viglet.shiohara.persistence.repository.folder.ShFolderRepository;
 import com.viglet.shiohara.persistence.repository.post.ShPostAttrRepository;
 import com.viglet.shiohara.persistence.repository.post.ShPostRepository;
+import com.viglet.shiohara.persistence.repository.post.type.ShPostTypeRepository;
 import com.viglet.shiohara.persistence.repository.reference.ShReferenceRepository;
+import com.viglet.shiohara.post.type.ShSystemPostType;
+import com.viglet.shiohara.turing.ShTuringIntegration;
 import com.viglet.shiohara.url.ShURLScheme;
 
 @Component
@@ -33,9 +40,23 @@ public class ShFolderUtils {
 	private ShPostAttrRepository shPostAttrRepository;
 	@Autowired
 	private ShReferenceRepository shReferenceRepository;
-
 	@Autowired
 	private ShURLScheme shURLScheme;
+	@Autowired
+	private ShTuringIntegration shTuringIntegration;
+	@Autowired
+	private ShPostTypeRepository shPostTypeRepository;
+
+	public ShPost getFolderIndex(ShFolder shFolder) {
+		ShPostType shPostType = shPostTypeRepository.findByName(ShSystemPostType.FOLDER_INDEX);
+		List<ShPost> shFolderIndexPosts = shPostRepository.findByShFolderAndShPostTypeOrderByPositionAsc(shFolder,
+				shPostType);
+		if (shFolderIndexPosts.size() > 0) {
+			ShPost shFolderIndexPost = shFolderIndexPosts.get(0);
+			return shFolderIndexPost;
+		}
+		return null;
+	}
 
 	public JSONObject toJSON(ShFolder shFolder) {
 		JSONObject shFolderItemAttrs = new JSONObject();
@@ -210,7 +231,8 @@ public class ShFolderUtils {
 	}
 
 	@Transactional
-	public boolean deleteFolder(ShFolder shFolder) {
+	public boolean deleteFolder(ShFolder shFolder) throws ClientProtocolException, IOException {
+		shTuringIntegration.deindexObject(shFolder);
 
 		for (ShPost shPost : shPostRepository.findByShFolder(shFolder)) {
 			// TODO: Check relation and show to user decides.
@@ -230,7 +252,9 @@ public class ShFolderUtils {
 		for (ShFolder shFolderChild : shFolderRepository.findByParentFolder(shFolder)) {
 			this.deleteFolder(shFolderChild);
 		}
+
 		shFolderRepository.delete(shFolder.getId());
+
 		return true;
 	}